requirements.txt
openai>=1.30
pyyaml>=6.0
markdown>=3.6
feedparser>=6.0
